remove obsolete seting `type` from example
The example raises this error when I use it with logstash 2.3.1: 

The setting `type` in plugin `multiline` is obsolete and is no longer available. You can achieve this same behavior with the new conditionals, like: `if [type] == \"sometype\" { multiline { ... } }`. If you have any questions about this, you are invited to visit https://discuss.elastic.co/c/logstash and ask.