Correctly handle reads by the input module that are not aligned to a newline
The input plugin might not align the reads by the newlines (i.e. stdin).
For example stdin reads 16384 bytes and passes them to the codec,
regardless if that read ends with a newline. This commit fixes
this by keeping the last 'partial' line and adds it with the next
decode call.

In our environment this change didn't have any noticable effect on the
performance.

Fixes Issue #37.